Move Registry, Restriction, Sanitize and Source class to namespace
